Merging Master into current_event_work

+++ volunteerquest/src/app/view-event/view-event.component.ts
@@ -6,6 +6,7 @@ import { Observable } from 'rxjs/Observable';
 import { Event } from '../manage-events/event.model';
 import { EventEditComponent } from '../event-edit/event-edit.component';
 import { EventManagerService } from '../services/search-engine/event-manager.service';
+import { ConfirmDeleteEventDialogComponent } from '../confirm-delete-event-dialog/confirm-delete-event-dialog.component';
 
 
 @Component({
@@ -41,9 +42,43 @@ export class ViewEventComponent implements OnInit {
     }
   }
 
-  deleteEvent(eventID: string) {
-    console.log("eventID:", eventID);
-    this.EventManagerService.remove(eventID);
+  // First confirms with user if they would like to delete the event and then if confirmed, 
+  // deletes the event
+  deleteEvent(eventID: string, eventContent: string, eventTitle: string, eventDate: Date, eventStreet: string) {
+    let dialogRef = this.dialog.open(ConfirmDeleteEventDialogComponent, {
+      width: '30em',
+      data: { eventContent: eventContent,
+            eventTitle: eventTitle,
+            eventDate: eventDate,
+            eventStreet: eventStreet }
+    });
+    
+    dialogRef.afterClosed().subscribe(result => {
+      console.log("The confirm delete event dialog was closed");
+      console.log("result:", result);
+      if (result) {
+        this.EventManagerService.remove(eventID);
+        console.log("event was deleted")
+      } else {
+        console.log("event was not deleted")
+      }
+    })
+  }
+
+  editEventDialog(event: Event) {
+    console.log("event:", JSON.stringify(event));
+    // Get the data associated with this event and pass it to the event dialog component
+    let dialogRef = this.dialog.open(EventEditComponent, {
+      width: '30em',
+      data: {event: event}
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      console.log("The edit event dialog was closed");
+      this.events$ = this.EventManagerService.getCollection$(ref => ref.where("uid", '==', this.uid));
+      if (!this.events$) {
+        console.log("no events");
+      }
+    })
   }
 
   openNewEventDialog() {
